+++ src/legacy/core_plugins/data/public/index.scss
@@ -1,4 +1,4 @@
 @import 'src/legacy/ui/public/styles/styling_constants';
 
-@import './query_bar/index';
+@import './query/query_bar/index';
