+++ src/legacy/core_plugins/data/public/query/query_bar/components/_query_bar.scss
