+++ src/legacy/core_plugins/data/public/query/query_bar/components/language_switcher.test.tsx
