+++ src/legacy/core_plugins/data/public/query/query_bar/components/query_bar.tsx
