+++ src/legacy/core_plugins/data/public/query/query_bar/components/query_bar_input.test.mocks.ts
@@ -18,7 +18,7 @@
  */
 
 import { createKfetch } from 'ui/kfetch/kfetch';
-import { setup } from '../../../../../../test_utils/public/kfetch_test_setup';
+import { setup } from 'test_utils/kfetch_test_setup';
 
 const mockChromeFactory = jest.fn(() => {
   return {
