+++ src/legacy/core_plugins/data/public/query/query_bar/components/typeahead/suggestion_component.tsx
