+++ src/legacy/core_plugins/data/public/query/query_bar/components/typeahead/suggestions_component.tsx
