+++ src/legacy/core_plugins/data/public/query/query_bar/directive/index.js
@@ -17,4 +17,22 @@
  * under the License.
  */
 
-export { SearchBarService } from './search_bar_service';
+import 'ngreact';
+import { wrapInI18nContext } from 'ui/i18n';
+import { uiModules } from 'ui/modules';
+import { QueryBar } from '../components';
+
+const app = uiModules.get('app/data', ['react']);
+
+export function setupDirective() {
+  app.directive('queryBar', (reactDirective, localStorage) => {
+    return reactDirective(
+      wrapInI18nContext(QueryBar),
+      undefined,
+      {},
+      {
+        store: localStorage,
+      }
+    );
+  });
+}
