+++ src/legacy/core_plugins/data/public/query/query_bar/lib/__tests__/lib_user_input.js
