+++ src/legacy/core_plugins/data/public/query/query_bar/lib/get_query_log.ts
