+++ src/legacy/core_plugins/data/public/query/query_bar/lib/match_pairs.ts
