+++ src/legacy/core_plugins/data/public/query/query_bar/lib/to_user.ts
