+++ src/legacy/core_plugins/data/public/search/search_bar/components/index.tsx
