+++ src/legacy/core_plugins/data/public/search/search_bar/directive/index.js
