+++ src/legacy/core_plugins/data/public/search/search_service.ts
@@ -17,24 +17,27 @@
  * under the License.
  */
 
+import { once } from 'lodash';
+import { SearchBar, setupDirective as setupSearchBarDirective } from './search_bar';
 
+/**
+ * Search Service
+ * @internal
+ */
+export class SearchService {
+  public setup() {
+    return {
+      ui: {
+        SearchBar,
+      },
+      loadLegacyDirectives: once(setupSearchBarDirective),
+    };
+  }
 
-import 'ngreact';
-import { wrapInI18nContext } from 'ui/i18n';
-import { uiModules } from 'ui/modules';
-import { QueryBar } from '../components';
-
-const app = uiModules.get('app/data', ['react']);
-
-export function setupDirective() {
-  app.directive('queryBar', (reactDirective, localStorage) => {
-    return reactDirective(
-      wrapInI18nContext(QueryBar),
-      undefined,
-      {},
-      {
-        store: localStorage,
-      }
-    );
-  });
+  public stop() {
+    // nothing to do here yet
+  }
 }
+
+/** @public */
+export type SearchSetup = ReturnType<SearchService['setup']>;
